refactor(date-input): replace TimeField with TimeSelect and update styles
- Updated test cases to use TimeSelect instead of TimeField in date-input tests.
- Refactored CSS variables and styles to reflect the change from TimeField to TimeSelect.
- Enhanced documentation to clarify the use of multiple date selection.
- Added new components for DayPickerPanel, MonthPickerPanel, PanelHeader, and YearPickerPanel to improve date picker functionality.
