+++ packages/varlet-ui/src/date-picker/MonthPickerPanel.vue
@@ -25,13 +25,13 @@
 <script lang="ts">
 import { defineComponent, ref } from 'vue'
 import type { PropType, Ref } from 'vue'
-import VarButton from '../../button'
-import { t } from '../../locale'
-import { injectLocaleProvider } from '../../locale-provider/provide'
-import { createNamespace } from '../../utils/components'
-import { createDayjs } from '../../utils/shared'
-import { DatePickerTypes, DatePickerUnits, MonthList, ShiftDirections, type Month } from '../constants'
-import type { DatePickerPreviewState, DatePickerSelectionState, PanelDatePickerProps } from '../types'
+import VarButton from '../button'
+import { t } from '../locale'
+import { injectLocaleProvider } from '../locale-provider/provide'
+import { createNamespace } from '../utils/components'
+import { createDayjs } from '../utils/shared'
+import { DatePickerTypes, DatePickerUnits, MonthList, ShiftDirections, type Month } from './constants'
+import type { DatePickerPreviewState, DatePickerSelectionState, PanelDatePickerProps } from './types'
 
 const dayjs = createDayjs()
 const { n } = createNamespace('month-picker')
