+++ packages/varlet-ui/src/date-picker/YearPickerPanel.vue
@@ -24,11 +24,11 @@
 import { times, toNumber } from '@varlet/shared'
 import { onSmartMounted } from '@varlet/use'
 import { computed, defineComponent, ref, watch, type ComputedRef, type PropType, type Ref } from 'vue'
-import VarButton from '../../button'
-import { createNamespace } from '../../utils/components'
-import { createDayjs } from '../../utils/shared'
-import { DatePickerTypes, DatePickerUnits, ShiftDirections } from '../constants'
-import { type DatePickerSelectionState, type PanelDatePickerProps } from '../types'
+import VarButton from '../button'
+import { createNamespace } from '../utils/components'
+import { createDayjs } from '../utils/shared'
+import { DatePickerTypes, DatePickerUnits, ShiftDirections } from './constants'
+import { type DatePickerSelectionState, type PanelDatePickerProps } from './types'
 
 const dayjs = createDayjs()
